+++ webextension/options/options.js
@@ -15,11 +15,11 @@ const box_notifications          = document.querySelector('input[name="disableNo
 const lbl_notifications          = document.querySelector('label[for="disableNotifications"]')
 const txt_server                 = document.querySelector('input[name="api_server"]')
 const lbl_server                 = document.querySelector('label[for="api_server"]')
-const txt_sha256                 = document.querySelector('input[name="api_sha256"]')
-const lbl_sha256                 = document.querySelector('label[for="api_sha256"]')
+const txt_publicKey              = document.querySelector('input[name="api_publicKey"]')
+const lbl_publicKey              = document.querySelector('label[for="api_publicKey"]')
 const btn_save                   = document.getElementById('form-submit')
 const btn_restoreDefault         = document.getElementById('restore-default')
-const btn_getFingerprints        = document.getElementById('get-fingerprints')
+const btn_getPublicKey           = document.getElementById('get-publicKey')
 const div_messageCheckServer     = document.querySelector('p.message-checkserver')
 
 
@@ -36,17 +36,17 @@ browser.runtime.getBackgroundPage().then((backgroundPage) => {
   lbl_alertIDNDomains.textContent      = browser.i18n.getMessage('__alertOnUnicodeIDNDomainNames__')
   lbl_notifications.textContent        = browser.i18n.getMessage('__disableNotifications__')
   lbl_server.textContent               = browser.i18n.getMessage('__checkServerAddress__')
-  lbl_sha256.textContent               = browser.i18n.getMessage('__checkServerSha256__')
+	lbl_publicKey.textContent               = browser.i18n.getMessage('__checkServerPublicKey__')
   btn_save.textContent                 = browser.i18n.getMessage('__save__')
   btn_restoreDefault.textContent       = browser.i18n.getMessage('__restoreDefault__')
-  btn_getFingerprints.textContent      = browser.i18n.getMessage('__getFingerprints__')
+  btn_getPublicKey.textContent      = browser.i18n.getMessage('__getPublicKey__')
 
   box_pageLoad.checked = CMH.options.settings.checkOnPageLoad
   box_alertIDNDomains.checked = CMH.options.settings.alertOnUnicodeIDNDomainNames
   box_notifications.checked = CMH.options.settings.disableNotifications
   lastDomainSaved  = CMH.options.settings.checkServerUrl.match(/^https:\/\/([^:\/\s]+)/)[1]
   txt_server.value = CMH.options.settings.checkServerUrl
-  txt_sha256.value = CMH.options.settings.checkServerFingerprintsSha256
+  txt_publicKey.value = CMH.options.settings.publicKey
 
   box_pageLoad.addEventListener('input', (e) => {
     browser.storage.local.set({
@@ -94,7 +94,7 @@ browser.runtime.getBackgroundPage().then((backgroundPage) => {
     const saveSettingsToBrowser = () => {
       browser.storage.local.set({
         checkServerUrl:                txt_server.value,
-        checkServerFingerprintsSha256: txt_sha256.value.replace(/:/g, '').toUpperCase()
+        publicKey: txt_publicKey.value
       }).then(() => {
         btn_save.disabled = false
         div_messageCheckServer.dataset.type = 'success'
@@ -107,16 +107,26 @@ browser.runtime.getBackgroundPage().then((backgroundPage) => {
     }
 
     if (CMH.common.isWebExtTlsApiSupported()) {
-      isValidCheckServer = await CMH.api.checkCheckServerApi({
-        server: txt_server.value,
-        sha256: txt_sha256.value.replace(/:/g, '').toUpperCase()
-      })
-      if (isValidCheckServer) {
+      isValidCheckServer = await CMH.options.verifyServerAtStartup(txt_server.value, txt_publicKey.value)
+      if (isValidCheckServer === 1) {
         saveSettingsToBrowser()
       } else {
         btn_save.disabled = false
         div_messageCheckServer.dataset.type = 'error'
-        div_messageCheckServer.textContent  = 'Error!'
+        switch (isValidCheckServer) {
+          case -1:
+            div_messageCheckServer.textContent  = browser.i18n.getMessage('__serverUnreachable__')
+            break;
+          case -2:
+            div_messageCheckServer.textContent  = browser.i18n.getMessage('__invalidPublicKeyInOptions__')
+            break;
+          case 0:
+            div_messageCheckServer.textContent  = browser.i18n.getMessage('__publicKeyNotCorresponding__')
+            break;
+          default:
+            div_messageCheckServer.textContent  = 'Error!'
+            break;
+        }
       }
     }
   }, true)
@@ -125,50 +135,47 @@ browser.runtime.getBackgroundPage().then((backgroundPage) => {
     const defaultCheckServer = CMH.options.defaultCheckServer
     if (defaultCheckServer !== null) {
       txt_server.value = defaultCheckServer.url
-      txt_sha256.value = defaultCheckServer.fingerprints.sha256
-      btn_getFingerprints.style.display = 'none'
+      txt_publicKey.value = defaultCheckServer.publicKey
+      btn_getPublicKey.style.display = 'none'
     }
   }, true)
 
-  btn_getFingerprints.addEventListener('click', (event) => {
-    btn_getFingerprints.disabled = true
+  btn_getPublicKey.addEventListener('click', (event) => {
+    btn_getPublicKey.disabled = true
     div_messageCheckServer.textContent = ''
     if (txt_server.value.slice(-1) !== '/') {
       txt_server.value += '/'
     }
 
-    CMH.options.getCertUrl(txt_server.value).then((response) => {
-      btn_getFingerprints.disabled = false
-      if (response.fingerprints !== null) {
-        txt_sha256.value = response.fingerprints.sha256
-        btn_getFingerprints.style.display = 'none'
+    CMH.certificatesManager.getCertUrl(txt_server.value+'download/public_key').then((response) => {
+      btn_getPublicKey.disabled = false
+      if (response.data !== null) {
+        txt_publicKey.value = response.data
+        btn_getPublicKey.style.display = 'none'
       } else {
         div_messageCheckServer.dataset.type = 'error'
-        div_messageCheckServer.textContent  = 'Error!'
+        div_messageCheckServer.textContent  = browser.i18n.getMessage('__publicKeyUnreachable__')
       }
     })
   }, true)
-
-  txt_sha256.addEventListener('blur', () => {
-    txt_sha256.value = txt_sha256.value.replace(/:/g, '').toUpperCase()
-  }, true)
-  const onFingerprintChange = () => {
-    if (txt_sha256.value.length === 0) {
-      btn_getFingerprints.style.display = ''
+  
+  const onPublicKeyChange = () => {
+    if (txt_publicKey.value.length === 0) {
+      btn_getPublicKey.style.display = ''
     } else {
-      btn_getFingerprints.style.display = 'none'
+      btn_getPublicKey.style.display = 'none'
     }
   }
-  txt_sha256.addEventListener('keyup', onFingerprintChange, true)
+  txt_publicKey.addEventListener('keyup', onPublicKeyChange, true)
   txt_server.addEventListener('keyup', () => {
     const domainMatch = txt_server.value.match(/^https:\/\/([^:\/\s]+)/)
     if (domainMatch && (domainMatch[1] !== lastDomainSaved)) {
-      txt_sha256.value = ''
-      btn_getFingerprints.style.display = ''
+      txt_publicKey.value = ''
+      btn_getPublicKey.style.display = ''
     }
   }, true)
 
-  onFingerprintChange()
+  onPublicKeyChange()
   btn_save.disabled = false
 
   document.body.style.display = ''